requirements.txt
-i https://pypi.org/simple
alembic==1.4.2
blinker==1.4
click==7.1.2
dnspython==2.0.0
dominate==2.5.2
email-validator==1.1.1
flask-bootstrap==3.3.7.1
flask-login==0.5.0
flask-mail==0.9.1
flask-migrate==2.5.3
flask-sqlalchemy==2.4.4
flask-wtf==0.14.3
flask==1.1.2
idna==2.10
itsdangerous==1.1.0
jinja2==2.11.2
mako==1.1.3
markupsafe==1.1.1
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
six==1.15.0
sqlalchemy==1.3.19
visitor==0.1.3
werkzeug==1.0.1
wtforms==2.3.3
